Conflicts:
	client/src/main/scala/events/EventsController.scala
	client/src/main/scala/httpServiceFactory/HttpGeneralService.scala
	client/src/main/scala/places/PlacesController.scala

+++ client/src/main/scala/places/PlacesRoutes.scala
@@ -2,7 +2,7 @@ package places
 
 object PlacesRoutes {
 
-  def findById(id: Long): String = "/places/" + id
+  def findById(id: Int): String = "/places/" + id
   
   def followByPlaceId(placeId: Long): String = "places/" + placeId + "/followByPlaceId"
